Restruct tests
