+++ testing/tests/Renovation/__meta.json
@@ -1,5 +1,5 @@
 {
-  "constellation": "viz",
+  "constellation": "renovation",
   "explicit": false,
   "runOnDevices": true
 }
